2024.7.25更新

+++ AestheticsLink/LogRegService/Dto/CustomerDto.cs
@@ -15,7 +15,7 @@ namespace LogRegService.Dto
         public string GENDER { get; set; }
         public string NAME { get; set; }
         public string PASSWORD { get; set; }
-        public int BALANCE { get; set; }
+        public decimal BALANCE { get; set; }
         public string VIPLEVEL { get; set; }
     }
 }
