+++ AestheticsLink/QueryAllUsersService/QueryAllCustomers/Dto/QueryCustomersDto.cs
@@ -15,7 +15,7 @@ namespace QueryAllUsersService.QueryAllCustomers.Dto
         public int age { get; set; }
         public DateTime birthday { get; set; }
         public string viplevel { get; set; }
-        public int balance { get; set; }
+        public decimal balance { get; set; }
     }
 
     public class DeleteCustomersDto
