+++ AestheticsLink/WebModel/Entity/BILL.cs
@@ -12,7 +12,7 @@ namespace WebModel.Entity
         public string CUS_ID { get; set; }
         public string COU_ID { get; set; }
         public string FOUND_DATE { get; set; }
-        public int PAID_AMOUNT { get; set; }
+        public decimal PAID_AMOUNT { get; set; }
         public string PAY_STATE { get; set; }
         public string HOS_ID { get; set; }
     }
